% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up_paypal.R
\name{clean_up_paypal}
\alias{clean_up_paypal}
\title{Bereinige und formatiere die rohen PayPal-Daten}
\usage{
clean_up_paypal(df)
}
\arguments{
\item{df}{Ein Data Frame mit den rohen PayPal-Daten.}
}
\value{
Ein bereinigter Data Frame (Tibble), optimiert für das Shiny Dashboard.
}
\description{
Wendet \code{clean_master()} an, reduziert die Spalten drastisch
(Whitelist-Ansatz), benennt sie lesbar um und standardisiert die
internationalen Ratenzahlungs-Bezeichnungen.
}
\examples{
\dontrun{
clean_paypal <- clean_up_paypal(raw_paypal_df)
}
}
