% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up_trustpilot.R
\name{clean_up_trustpilot}
\alias{clean_up_trustpilot}
\title{Clean Trustpilot Data}
\usage{
clean_up_trustpilot(df)
}
\arguments{
\item{df}{Data.frame oder Tibble. Die rohen Trustpilot-Daten aus dem API-Call.}
}
\value{
Ein bereinigtes `tibble` mit den relevanten KPIs.
}
\description{
Bereinigt die rohen Trustpilot-API-Daten durch Entfernen unnötiger
Link- und Metadaten-Spalten sowie durch die Formatierung von Datums- und Zahlenwerten.
Nutzt im ersten Schritt die interne Funktion `clean_master()`, um die Spaltennamen
zu standardisieren.
}
\examples{
\dontrun{
tp_clean <- clean_up_trustpilot(trustpilot_raw)
}
}
