% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_zendesk_csv_tickets.R
\name{clean_zendesk_csv_tickets}
\alias{clean_zendesk_csv_tickets}
\title{Clean and Map Zendesk Ticket Data}
\usage{
clean_zendesk_csv_tickets(
  data,
  save = TRUE,
  save_path = file.path(datadir, "zendesk/zendesk_csv_tickets_cleaned.rds")
)
}
\arguments{
\item{data}{A data.frame containing Zendesk ticket data.}

\item{save}{Logical. Whether to save the cleaned data as an RDS file. Default is \code{TRUE}.}

\item{save_path}{Character. Full file path for saving the cleaned data. Default is \code{file.path(datadir, "zendesk/zendesk_csv_tickets_cleaned.rds")}.}

\item{datadir}{Character. Directory path where the cleaned file should be saved (if \code{save = TRUE}).}
}
\value{
The cleaned data.frame with mapped assignee and group names.
}
\description{
Cleans a Zendesk tickets data.frame by removing unnecessary columns, mapping assignee and group IDs to names, and optionally saving the cleaned data.
}
