% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paypal_token.R
\name{get_paypal_token}
\alias{get_paypal_token}
\title{Request PayPal OAuth 2.0 Access Token}
\usage{
get_paypal_token(
  client_id = Sys.getenv("PAYPAL_CLIENT_ID"),
  secret = Sys.getenv("PAYPAL_SECRET"),
  base_url = Sys.getenv("PAYPAL_BASE_URL")
)
}
\arguments{
\item{client_id}{Character. The PayPal Client ID. Defaults to the environment variable `PAYPAL_CLIENT_ID`.}

\item{secret}{Character. The PayPal Secret. Defaults to the environment variable `PAYPAL_SECRET`.}

\item{base_url}{Character. The PayPal API base URL. Defaults to the environment variable `PAYPAL_BASE_URL`.}
}
\value{
Character string containing the access token.
}
\description{
Authenticates with the PayPal REST API using Client Credentials
stored in the `.Renviron` file and retrieves a short-lived Bearer access token.
}
\examples{
\dontrun{
token <- get_paypal_token()
}
}
