% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_adtribute_master.R
\name{load_adtribute_master}
\alias{load_adtribute_master}
\title{Initialisiert den virtuellen Master-Datensatz (Lazy Evaluation)}
\usage{
load_adtribute_master(local_r_path = "~/workspace/local.R")
}
\arguments{
\item{local_r_path}{Pfad zur lokalen Konfigurationsdatei. Standard: "~/workspace/local.R"}
}
\value{
Ein virtuelles Arrow-Dataset/Query-Objekt (master_lazy)
}
\description{
Initialisiert den virtuellen Master-Datensatz (Lazy Evaluation)
}
