% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lottery.R
\name{lottery}
\alias{lottery}
\title{Ziehung der Gewinner eines Gewinnspiels}
\usage{
lottery(data, prizes = c(rep("price1", 10), rep("price2", 100)), seed = NULL)
}
\arguments{
\item{data}{Ein Dataframe mit mindestens den Spalten: id, email, first_name, last_name, amount_lose.}

\item{prizes}{Ein Vektor mit den zu vergebenden Preisen (z.B. von \code{rep()} erzeugt). Default: vordefinierte Preise.}

\item{seed}{Optional. Zufalls-Seed für Reproduzierbarkeit. Default ist NULL (kein gesetzter Seed).}
}
\value{
Ein Dataframe mit den Gewinnern und den zugewiesenen Preisen.
}
\description{
Diese Funktion führt eine gewichtete Ziehung von Gewinnern anhand der Anzahl der Lose durch
und weist zufällig Preise zu.
}
\examples{
winners <- lottery(gewinnspiel_filtered, seed = 123)
}
