% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dashboardly_data.R
\name{get_dashboardly_data}
\alias{get_dashboardly_data}
\title{Fetch All Orders from Dashboardly API}
\usage{
get_dashboardly_data(
  shop_id,
  start_date,
  end_date,
  api_key = Sys.getenv("DASHBOARDLY_API_KEY")
)
}
\arguments{
\item{shop_id}{character. Die ID des Shops (z. B. "3d2ec700-...").}

\item{start_date}{character. Startdatum im Format "YYYY-MM-DD".}

\item{end_date}{character. Enddatum im Format "YYYY-MM-DD".}

\item{api_key}{character. Dashboardly API Key. Standardmäßig via `.Renviron`.}
}
\value{
Ein Tibble mit allen Bestellungen des angegebenen Zeitraums.
}
\description{
Holt alle paginierten Bestelldaten für einen spezifischen Shop
  über die Dashboardly Developer API. Iteriert automatisch durch alle Seiten.
}
\examples{
\dontrun{
tiktok_orders <- get_dashboardly_orders(
  shop_id = "3d2ec700-f2a3-468b-91ba-ec8826b6ab43",
  start_date = "2026-07-01",
  end_date = "2026-07-08"
)
}
}
