% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shopify_data.R
\name{get_shopify_data}
\alias{get_shopify_data}
\title{Fetch Data from Shopify API (with Pagination and Checkpointing)}
\usage{
get_shopify_data(
  shop_url = "pummmys.myshopify.com",
  api_key,
  endpoint = "orders",
  api_version = "2024-01",
  limit = 250,
  chunk_size = 100,
  temp_dir = "data/shopify_temp"
)
}
\arguments{
\item{shop_url}{Character. Die Shopify-Shop-URL.}

\item{api_key}{Character. Das Shopify Admin API Access Token.}

\item{endpoint}{Character. Der gewuenschte API-Endpunkt.}

\item{api_version}{Character. Die zu verwendende Shopify API-Version.}

\item{limit}{Integer. Anzahl der Eintraege pro Seite (Max: 250).}

\item{chunk_size}{Integer. Nach wie vielen Seiten soll ein Zwischenstand auf der Festplatte gespeichert werden?}

\item{temp_dir}{Character. Pfad zum Verzeichnis, in dem die Chunks gespeichert werden.}
}
\value{
Ein tibble mit allen geparsten Daten.
}
\description{
Ruft Daten von einem spezifizierten Shopify API-Endpunkt ab.
Nutzt Chunking, um bei grossen Datenmengen (z. B. > 100.000 Bestellungen) den Arbeitsspeicher
zu schonen. Speichert Zwischenstaende als .rds-Dateien.
}
