% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_bigquery.R
\name{upload_to_bigquery}
\alias{upload_to_bigquery}
\title{Upload Data to BigQuery}
\usage{
upload_to_bigquery(
  data,
  dataset_id = "support_analytics",
  table_id,
  json_key_path = "musterkeypath",
  write_disposition = "WRITE_APPEND"
)
}
\arguments{
\item{data}{Ein data.frame oder tibble, das hochgeladen werden soll.}

\item{dataset_id}{Character. Die BigQuery Dataset-ID.}

\item{table_id}{Character. Der Name der Zieltabelle (wird erstellt, falls nicht existent).}

\item{json_key_path}{Character. Der lokale Pfad zur .json Schlüsseldatei.}

\item{write_disposition}{Character. Verhalten, wenn die Tabelle schon existiert. Standard ist "WRITE_APPEND". Alternativ: "WRITE_TRUNCATE".}
}
\value{
Invisible TRUE bei erfolgreichem Upload.
}
\description{
Authentifiziert sich via JSON-Service-Account-Key, extrahiert automatisch die Projekt-ID und lädt ein Dataframe in eine BigQuery Tabelle hoch.
}
\details{
Upload data to Google BigQuery
}
\examples{
\dontrun{
upload_to_bigquery(
  data = clean_zendesk_data,
  table_id = "zendesk_tickets"
)
}
}
