% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_bigquery.R
\name{upload_to_bigquery}
\alias{upload_to_bigquery}
\title{Upload Data to BigQuery}
\usage{
upload_to_bigquery(
  data,
  dataset_id = "support_analytics",
  table_id,
  json_key_path = "~/git/dieseoR/scripts/data-analytics-491117-58f718e1ee61.json",
  write_disposition = "WRITE_APPEND"
)
}
\description{
Authentifiziert sich und lädt via ausfallsicherem Parquet-Format nach BigQuery hoch.
}
